breaking out transformations into their own objects in the transformations package
